+++ CHANGELOG.md
@@ -4,6 +4,32 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Fix truncated title styles on `CrowdfundingCard`.
+
+## [20.0.1] - 2018-03-05
+
+Fixes:
+- Add missing `event` parameter to `Pagination`'s onClick.
+- Update `font-size` big on `TagButton` component.
+- Update Radium.
+
+## [20.0.0] - 2018-03-01
+
+Breaking changes:
+- Rename `style` prop to `fontStyle` in `Text` component.
+- You must now add a `<StyleRoot>` around `<Hero>` components.
+
+Features:
+- Add `Pagination` component.
+- Update styles for `disabled` prop on `ArrowIcon` component.
+- Add `rampProps` prop to `Progress` component.
+- Add `CrowdfundingCard` component.
+- Add `mediaQueries` HOC.
+- Add `modifier` prop on `TagButton` component.
+
+Fix:
+- Update `primary-light` color.
+
 ## [19.5.0] - 2018-02-13
 
 Feature: